Refactoring version generation with type class implicits, using implicitly and a companion object to get the predef implicits in place without impoting.
